requirements.txt
pyautogui==0.9.53
keyboard==0.13.5
numpy==1.21.2
pillow==8.4.0
pywin32==303
scikit-image
SQLalchemy